backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
authlib==1.3.0
itsdangerous==2.1.2
pydantic[email]==2.5.3
pydantic-settings==2.1.0
